website/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-filter==21.1
Pillow==8.4.0
python-dotenv=0.19.2